+++ src/TypeUtils.jl
@@ -22,6 +22,7 @@ export
     destructure!,
     destructure,
     floating_point_type,
+    new_array,
     parameterless,
     promote_eltype,
     real_type,
@@ -86,6 +87,34 @@ See also [`ArrayAxes`](@ref), `Dims`.
 
 """
 const ArrayShape{N} = NTuple{N,Union{Integer,AbstractUnitRange{<:Integer}}}
+
+"""
+    TypeUtils.Unsupported(T::DataType...)
+
+yields an union of types `T...` and of type `Unsupported`. Such an union can be used to
+mark unsupported argument type(s) and yet define a method applicable with that type(s)
+(presumably a method that throws an instructive error) and which can be extended later
+with the same signature except that with `Unsupported(T...)` replaced by `T...`. This
+trick avoids conflicts that prevent pre-compilation with package extensions.
+
+For example, in the main package:
+
+```julia
+some_method(some_arg::Unsupported{SomeType}) =
+    error("package `SomeOtherPackage` has not yet been loaded")
+```
+
+and in the extension (e.g. automatically loaded when using `SomeOtherPackage`):
+
+```julia
+some_method(some_arg::SomeType) = do_something_with(some_arg)
+```
+
+"""
+struct Unsupported
+    Unsupported() = error("it is not possible to instanciate this type")
+end
+Unsupported(T::DataType...) = Union{T...,Unsupported}
 
 """
     as(T, x)
@@ -216,8 +245,8 @@ converts array dimensions or ranges `args...` to a canonical form of array shape
 The array dimensions or ranges may also be provided as a tuple. [`ArrayShape{N}`](@ref) is
 the union of types of `N`-tuples to which `as_array_shape` is applicable.
 
-Also see [`as_array_size`](@ref), [`as_array_axes`](@ref), [`ArrayAxes`](@ref), and
-`Dims`.
+Also see [`as_array_size`](@ref), [`as_array_axes`](@ref), [`ArrayAxes`](@ref), `Dims`,
+and [`new_array`](@ref).
 
 """
 as_array_shape(::Tuple{}) = ()
@@ -234,7 +263,8 @@ tuple of `eltype(Dims)`s. Any range in `args...` is replaced by its length.
 The array dimensions or ranges may also be provided as a tuple. [`ArrayShape{N}`](@ref) is
 the union of types of `N`-tuples to which `as_array_size` is applicable.
 
-Also see [`as_array_shape`](@ref), [`as_array_axes`](@ref), [`as_array_dim`](@ref), and `Dims`.
+Also see [`as_array_shape`](@ref), [`as_array_axes`](@ref), [`as_array_dim`](@ref),
+`Dims`, and [`new_array`](@ref).
 
 """
 as_array_size(::Tuple{}) = ()
@@ -253,7 +283,7 @@ The array dimensions or ranges may also be provided as a tuple. [`ArrayShape{N}`
 the union of types of `N`-tuples to which `as_array_axes` is applicable.
 
 Also see [`as_array_shape`](@ref), [`as_array_size`](@ref), [`as_array_axis`](@ref),
-[`ArrayAxes`](@ref), and `Dims`.
+[`ArrayAxes`](@ref), `Dims`, and [`new_array`](@ref).
 
 """
 as_array_axes(::Tuple{}) = ()
@@ -287,6 +317,25 @@ Also see [`as_array_axes`](@ref), [`as_array_dim`](@ref), and `Dims`.
 as_array_axis(dim::Integer) = Base.OneTo{Dim}(dim)
 as_array_axis(rng::ArrayAxis) = rng
 as_array_axis(rng::AbstractUnitRange{<:Integer}) = as(ArrayAxis, rng)
+
+"""
+    new_array(T, inds...) -> A
+    new_array(T, (inds...,)) -> A
+
+create a new array with element type `T` and shape defined by `inds...`, a list of array
+dimension lengths and/or index ranges. The shape may also be specified as a tuple.
+
+If `inds...` contains any index range other than `Base.OneTo`, an `OffsetArray{T}` is
+returned; otherwise an `Array{T}` is returned. In the former case, an exception is thrown
+if the package `OffsetArrays` has not been loaded.
+
+"""
+new_array(::Type{T}, inds::eltype(ArrayShape)...) where {T} = new_array(T, inds)
+new_array(::Type{T}, inds::ArrayShape{N}) where {T,N} = new_array(T, as_array_shape(inds))
+new_array(::Type{T}, rngs::NTuple{N,Base.OneTo}) where {T,N} = new_array(T, as_array_size(rngs))
+new_array(::Type{T}, dims::Dims{N}) where {T,N} = Array{T,N}(undef, dims)
+new_array(::Type{T}, rngs::Unsupported(ArrayAxes{N})) where {T,N} =
+    error("package `OffsetArrays` must be loaded for such array index ranges")
 
 """
     return_type(f, argtypes...) -> T
@@ -914,9 +963,11 @@ end
 
 function __init__()
     @static if !isdefined(Base, :get_extension)
-        # Extend methods to `Unitful` quantities when this package is loaded.
+        # Extend methods when other packages are loaded.
         @require Unitful = "1986cc42-f94f-5a68-af5c-568840ba703d" include(
             "../ext/TypeUtilsUnitfulExt.jl")
+        @require OffsetArrays = "6fe1bfb0-de20-5000-8ca7-80f57d26f881" include(
+            "../ext/TypeUtilsOffsetArraysExt.jl")
     end
 end